Refactor commands to packages
