+++ src/main/java/seedu/duke/commands/food/AddFoodCommand.java
@@ -1,5 +1,6 @@
-package seedu.duke.commands;
+package seedu.duke.commands.food;
 
+import seedu.duke.commands.Command;
 import seedu.duke.exceptions.ArgumentsNotFoundException;
 import seedu.duke.food.FoodRecord;
 import seedu.duke.exceptions.IllegalFoodParameterException;
