+++ src/main/java/seedu/duke/commands/food/DeleteFoodCommand.java
@@ -1,5 +1,6 @@
-package seedu.duke.commands;
+package seedu.duke.commands.food;
 
+import seedu.duke.commands.Command;
 import seedu.duke.exceptions.FoodIndexNotFoundException;
 import seedu.duke.food.FoodRecord;
 import seedu.duke.parser.Parser;
