+++ src/main/java/seedu/duke/commands/food/ListFoodCommand.java
@@ -1,5 +1,6 @@
-package seedu.duke.commands;
+package seedu.duke.commands.food;
 
+import seedu.duke.commands.Command;
 import seedu.duke.storage.Storage;
 import seedu.duke.task.TaskList;
 import seedu.duke.ui.Ui;
@@ -10,7 +11,7 @@ import java.util.Scanner;
  * List all current records (assumed today for simplicity).
  * @author ngnigel99
  */
-public class ListFoodCommand extends Command  {
+public class ListFoodCommand extends Command {
 
     public ListFoodCommand() {
         this.helpMessage = "Lists all food items recorded";
