+++ src/main/java/seedu/duke/commands/module/DeleteModuleCommand.java
@@ -1,5 +1,6 @@
-package seedu.duke.commands;
+package seedu.duke.commands.module;
 
+import seedu.duke.commands.Command;
 import seedu.duke.exceptions.ClickException;
 import seedu.duke.exceptions.IllegalModuleIndexException;
 import seedu.duke.module.ModuleList;
