+++ src/main/java/seedu/duke/parser/Parser.java
@@ -1,24 +1,24 @@
 package seedu.duke.parser;
 
-import seedu.duke.commands.AddEntryCommand;
-import seedu.duke.commands.AddFoodCommand;
-import seedu.duke.commands.AddModuleCommand;
-import seedu.duke.commands.AddNoteCommand;
-import seedu.duke.commands.AddTodoCommand;
-import seedu.duke.commands.AddZoomCommand;
-import seedu.duke.commands.ClearFoodCommand;
+import seedu.duke.commands.journal.AddEntryCommand;
+import seedu.duke.commands.food.AddFoodCommand;
+import seedu.duke.commands.module.AddModuleCommand;
+import seedu.duke.commands.journal.AddNoteCommand;
+import seedu.duke.commands.calendar.AddTodoCommand;
+import seedu.duke.commands.zoom.AddZoomCommand;
+import seedu.duke.commands.food.ClearFoodCommand;
 import seedu.duke.commands.Command;
-import seedu.duke.commands.DeleteFoodCommand;
-import seedu.duke.commands.DeleteModuleCommand;
-import seedu.duke.commands.DeleteTaskCommand;
-import seedu.duke.commands.DisplayCalendarCommand;
+import seedu.duke.commands.food.DeleteFoodCommand;
+import seedu.duke.commands.module.DeleteModuleCommand;
+import seedu.duke.commands.calendar.DeleteTaskCommand;
+import seedu.duke.commands.calendar.DisplayCalendarCommand;
 import seedu.duke.commands.ExitCommand;
 import seedu.duke.commands.HelpCommand;
-import seedu.duke.commands.ListFoodCommand;
-import seedu.duke.commands.ListJournalCommand;
-import seedu.duke.commands.ListModuleCommand;
-import seedu.duke.commands.ListTasksCommand;
-import seedu.duke.commands.ShowZoomLinks;
+import seedu.duke.commands.food.ListFoodCommand;
+import seedu.duke.commands.journal.ListJournalCommand;
+import seedu.duke.commands.module.ListModuleCommand;
+import seedu.duke.commands.calendar.ListTasksCommand;
+import seedu.duke.commands.zoom.ShowZoomLinks;
 import seedu.duke.exceptions.IncorrectNumberOfArgumentsException;
 import seedu.duke.exceptions.ArgumentsNotFoundException;
 import seedu.duke.exceptions.ClickException;
