+++ src/test/java/seedu/duke/commands/AddTodoCommandTest.java
@@ -2,6 +2,7 @@ package seedu.duke.commands;
 
 import org.junit.jupiter.api.Test;
 
+import seedu.duke.commands.calendar.AddTodoCommand;
 import seedu.duke.exceptions.InvalidDateException;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
