+++ src/test/java/seedu/duke/commands/DisplayCalendarCommandTest.java
@@ -1,6 +1,7 @@
 package seedu.duke.commands;
 
 import org.junit.jupiter.api.Test;
+import seedu.duke.commands.calendar.DisplayCalendarCommand;
 import seedu.duke.exceptions.InvalidMonthException;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
