+++ src/test/java/seedu/duke/food/FoodTest.java
@@ -1,9 +1,7 @@
 package seedu.duke.food;
 
-import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
-import seedu.duke.commands.AddFoodCommand;
 import seedu.duke.constants.Messages;
 import seedu.duke.exceptions.ArgumentsNotFoundException;
 import seedu.duke.exceptions.IllegalFoodParameterException;
